Fix library inclusion for the r20 tools.
Instead of simply adding the android.library.reference to the user
properties list, this change actually inflates a project.properties
template to list the libraries. This allows the implementation of
library to change so long as the file format stays the same.

Prior to this, attrs used in <declare-styleable> were not being
referenced correctly.